feat(node): retry a failed proxy http request once
We observed random overlay nodes receiving ineligible responses from
data proxies while the overlay node should have been eligible. We
suspect the problem is that not all nodes have the new chain state, so
while the node that the overlay is talking to confirms that the overlay
is eligible, the node that the data proxy talks to is not yet at that
state. We now retry after a second as a simple bandaid.

Closes: #72